requirements.txt
gradio
torch
torchvision
transformers
pillow
requests
opencv-python
numpy
scipy
matplotlib
